referral_system/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
